Refactor - #6445 - CheckingInfoTool
