+++ src/components/CheckInfoCardWrapper.js
@@ -3,62 +3,23 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { CheckInfoCard } from 'tc-ui-toolkit';
 import { VerseObjectUtils } from 'word-aligner';
+// helpers
 import { TRANSLATION_NOTES, TRANSLATION_WORDS } from '../helpers/consts';
+import { getPhraseFromTw, getNote } from '../helpers/checkingInfoHelpers';
 
-class CheckInfoCardWrapper extends React.Component {
-  constructor(props) {
-    super(props);
-    this.getScriptureFromReference = this.getScriptureFromReference.bind(this);
-  }
-
-  getPhraseFromTw(translationWords, articleId, translationHelps) {
-    let currentFile = '';
-
-    if (translationWords && translationWords[articleId]) {
-      currentFile = translationHelps.translationWords[articleId];
-    }
-
-    let splitLine = currentFile.split('\n');
-
-    if (splitLine.length === 1 && splitLine[0] === '') {
-      return '';
-    }
-
-    let finalArray = [];
-
-    for (let i = 0; i < splitLine.length; i++) {
-      if (splitLine[i] !== '' && !~splitLine[i].indexOf('#')) {
-        finalArray.push(splitLine[i]);
-      }
-    }
-
-    let maxLength = 225;
-    let finalString = '';
-    let chosenString = finalArray[0];
-    let splitString = chosenString.split(' ');
-
-    for (let word of splitString) {
-      if ((finalString + ' ' + word).length >= maxLength) {
-        finalString += '...';
-        break;
-      }
-      finalString += ' ';
-      finalString += word;
-    }
-    return finalString;
-  }
-
-  getNote(occurrenceNote) {
-    // Removes the (See: [...](rc://...)) links at the end of an occurrenceNote
-    // Ex: Paul speaks of God’s message as if it were an object (See: [Idiom](rc://en/ta/man/translate/figs-idiom) and [Metaphor](rc://en/ta/man/translate/figs-metaphor)) =>
-    //     Paul speaks of God’s message as if it were an object
-    return occurrenceNote.replace(/\s*\([^\(\)\[\]]+((\[[^\[\]]+\])*(\[\[|\()+rc:\/\/[^\)\]]+(\]\]|\))[^\(\[\)\]]*)+[^\(\)\[\]]*\)\s*$/g, '');
-  }
-
-  getScriptureFromReference(lang, id, book, chapter, verse) {
+function CheckInfoCardWrapper({
+  translate,
+  translationHelps,
+  groupsIndex,
+  contextId,
+  resourcesReducer,
+  showHelps,
+  toggleHelps,
+}) {
+  function getScriptureFromReference(lang, id, book, chapter, verse) {
     const chapterParsed = parseInt(chapter);
     const verseParsed = parseInt(verse);
-    const currentBible = this.props.resourcesReducer.bibles[lang];
+    const currentBible = resourcesReducer.bibles[lang];
 
     if (currentBible &&
       currentBible[id] &&
@@ -70,46 +31,35 @@ class CheckInfoCardWrapper extends React.Component {
     }
   }
 
-  render() {
-    const {
-      translate,
-      translationHelps,
-      groupsIndex,
-      contextId,
-      showHelps,
-      toggleHelps,
-    } = this.props;
+  const {
+    groupId, occurrenceNote, tool,
+  } = contextId;
+  const title = groupsIndex.filter(item => item.id === groupId)[0].name;
+  let phrase = '';
 
-    const {
-      groupId, occurrenceNote, tool,
-    } = contextId;
-    const title = groupsIndex.filter(item => item.id === groupId)[0].name;
-    let phrase = '';
-
-    switch (tool) {
-    case TRANSLATION_WORDS: {
-      const { translationWords } = translationHelps ? translationHelps : {};
-      phrase = this.getPhraseFromTw(translationWords, contextId.groupId, translationHelps);
-      break;
-    }
-    case TRANSLATION_NOTES:
-      phrase = this.getNote(occurrenceNote);
-      break;
-    default:
-      console.error('tool is undefined in contextId');
-      break;
-    }
-
-    return (
-      <CheckInfoCard
-        title={title}
-        phrase={phrase}
-        getScriptureFromReference={this.getScriptureFromReference}
-        seeMoreLabel={translate('see_more')}
-        showSeeMoreButton={!showHelps}
-        onSeeMoreClick={toggleHelps} />
-    );
+  switch (tool) {
+  case TRANSLATION_WORDS: {
+    const { translationWords } = translationHelps ? translationHelps : {};
+    phrase = getPhraseFromTw(translationWords, contextId.groupId, translationHelps);
+    break;
   }
+  case TRANSLATION_NOTES:
+    phrase = getNote(occurrenceNote);
+    break;
+  default:
+    console.error('tool is undefined in contextId');
+    break;
+  }
+
+  return (
+    <CheckInfoCard
+      title={title}
+      phrase={phrase}
+      getScriptureFromReference={getScriptureFromReference}
+      seeMoreLabel={translate('see_more')}
+      showSeeMoreButton={!showHelps}
+      onSeeMoreClick={toggleHelps} />
+  );
 }
 
 CheckInfoCardWrapper.propTypes = {
